unify chat model in `websocket_wiki` and `simple_chat`
